refactor: NavHostViewModel stage 2 - extract banner and version check state holders
